+++ addons/web/static/src/core/commands/command_category.js
