+++ addons/web/static/src/core/commands/command_palette.scss
