+++ addons/web/static/src/core/commands/default_providers.js
