+++ addons/web/static/tests/core/commands/command_palette_tests.js
@@ -2,12 +2,12 @@
 
 import { makeDeferred } from "@mail/utils/deferred/deferred";
 import { browser } from "@web/core/browser/browser";
+import { CommandPaletteDialog } from "@web/core/commands/command_palette_dialog";
+import { commandService } from "@web/core/commands/command_service";
 import { dialogService } from "@web/core/dialog/dialog_service";
 import { hotkeyService } from "@web/core/hotkeys/hotkey_service";
 import { registry } from "@web/core/registry";
 import { uiService } from "@web/core/ui/ui_service";
-import { CommandPaletteDialog } from "@web/webclient/commands/command_palette_dialog";
-import { commandService } from "@web/webclient/commands/command_service";
 import { makeTestEnv } from "../../helpers/mock_env";
 import { click, getFixture, nextTick, patchWithCleanup, triggerHotkey } from "../../helpers/utils";
 import { editSearchBar } from "./command_service_tests";
