+++ addons/web/static/tests/core/commands/menu_provider_tests.js
