+++ addons/web/static/tests/webclient/navbar_tests.js
@@ -2,7 +2,7 @@
 
 import { browser } from "@web/core/browser/browser";
 import { notificationService } from "@web/core/notifications/notification_service";
-import { menuService } from "@web/webclient/menu_service";
+import { menuService } from "@web/webclient/menus/menu_service";
 import { registry } from "@web/core/registry";
 import { uiService } from "@web/core/ui/ui_service";
 import { actionService } from "@web/webclient/actions/action_service";
